+++ hnote/actions/index.ts
@@ -17,12 +17,23 @@ export const addTaskToList = (listId: string, str: string) => {
     };
 };
 
-export const updateTask = (itemId: string, updatedValue: string) => {
+export const updateTask = (itemId: string, updatedValue: any) => {
     return {
         type: 'UPDATE_TASK',
         payload: {
             itemID: itemId,
             content: updatedValue,
+        },
+    };
+};
+
+export const updateTaskIsFinished = (itemId: string, active: boolean) => {
+    console.log(!active);
+    return {
+        type: 'UPDATE_TASK_IS_FINISHED',
+        payload: {
+            itemID: itemId,
+            isDone: !active,
         },
     };
 };