+++ hnote/components/atoms/Texts/Paragraph.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+import { Text } from 'react-native';
+
+const Paragraph = styled.Text`
+    font-size: 20px;
+    font-weight: 300;
+    padding: 0;
+    margin: 0 10px;
+    text-decoration: ${({ isFinished }) =>
+        isFinished ? 'line-through' : 'none'};
+`;
+
+export default Paragraph;